Move calendar filters into a specific component
Easier to manager all the custom logic from here
