+++ addon/components/dashboard-calendar.hbs
@@ -237,38 +237,21 @@
       {{/liquid-if}}
     </div>
   </section>
-  {{#if this.activeFilters}}
-    <section
-      class="filters-list"
-      {{did-insert (perform this.loadFilterTags) this.cohortProxies this.activeFilters}}
-      {{did-update (perform this.loadFilterTags) this.cohortProxies this.activeFilters}}
-    >
-      {{#if this.filterTags}}
-        <header class="filters-header">
-          {{t "general.activeFilters"}}:
-        </header>
-        <div class="filter-tags">
-          {{#each (sort-by "name" this.filterTags) as |tag|}}
-            <span
-              class="filter-tag {{tag.class}}"
-              role="button"
-              {{on "click" (fn tag.remove tag.id)}}
-            >
-              {{tag.name}}
-              <FaIcon @icon="times" />
-            </span>
-          {{/each}}
-          <span
-            id="calendar-clear-filters"
-            class="filters-clear-filters"
-            role="button"
-            {{on "click" @clearFilters}}
-          >
-            {{t "general.clearFilters"}}
-          </span>
-        </div>
-      {{/if}}
-    </section>
+  {{#if (is-array this.cohortProxies)}}
+    <Dashboard::FilterTags
+      @selectedCourseLevels={{@selectedCourseLevels}}
+      @selectedSessionTypeIds={{@selectedSessionTypeIds}}
+      @selectedCohortIds={{@selectedCohortIds}}
+      @selectedCourseIds={{@selectedCourseIds}}
+      @selectedTermIds={{@selectedTermIds}}
+      @cohortProxies={{this.cohortProxies}}
+      @removeCourseLevel={{@removeCourseLevel}}
+      @removeSessionTypeId={{@removeSessionTypeId}}
+      @removeCohortId={{@removeCohortId}}
+      @removeCourseId={{@removeCourseId}}
+      @removeTermId={{@removeTermId}}
+      @clearFilters={{@clearFilters}}
+    />
   {{/if}}
   <section class="fullwidth ilios-calendar-container"
     {{did-insert (perform this.loadEvents) this.bestSelectedSchool this.fromTimeStamp this.toTimeStamp @mySchedule}}
