+++ app/components/dashboard/filter-tags.js
@@ -1,0 +1,1 @@
+export { default } from 'ilios-common/components/dashboard/filter-tags';
